refactor: move logic to hooks in tileset component
